refactor: moved /infraction_points into moderation.py
